service/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-decouple==3.8
gunicorn==21.2.0
django-redis==5.2.0